Move initializeJmxSeedsForAllClusters into constructor of JmxManagementConnectionFactory.
